pyproject.toml
[project]
name = "videre"
version = "0.1.0"
description = "Simple GUI framework based on Pygame"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cursword>=0.1.1",
    "filedial>=0.1.0",
    "fonttools>=4.59.1",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "pygame>=2.6.1",
    "pyperclip>=1.9.0",
    "tqdm>=4.67.1",
    "unicodedataplus>=16.0.0.post1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-regressions>=2.8.2",
]

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"
